requirements.txt
pyyaml>=6.0.0
lxml>=5.0.0
